+++ app/src/application/usecase/attackAlarm.ts
@@ -1,11 +1,10 @@
 import { LineNotifyService } from "../services/coc/lineNotifyService";
 import { BandService } from "../services/coc/band/bandService";
-import { CocApi } from "@src/infrastructure/http/cocApi";
-import { CurrentWar } from "@src/domain/currentWar/CurrentWar";
+import { CocApi } from "@src/infrastructure/http/cocApi/cocApi";
 import { ClanStoreRepository } from "@src/application/repository/ClanStoreRepository";
 import { LineNotify } from "@src/infrastructure/http/line/lineNotifyApi";
 import { ClanTag } from "@src/domain/ClanTag";
-import { ClanWarService } from "../services/coc/ClanWar";
+import { ClanWarService } from "../services/coc/clanWar/ClanWar";
 
 export class AttackAlarm {
     constructor(
@@ -15,6 +14,7 @@ export class AttackAlarm {
         private cocApi: CocApi,
         private lineNotufyService: LineNotifyService
     ) {}
+
     toLine = async (
         clanTag: ClanTag,
         alertHours: number[] = [1, 3, 6, 12, 24]
@@ -41,16 +41,12 @@ export class AttackAlarm {
     refreshPost = async (clanTag: ClanTag) => {
         const clan = await this.clanStoreRepository.getByTag(clanTag);
         if (!clan.band) return;
-        const currentWar = new CurrentWar(
-            await this.cocApi.getClanWarByTag(clanTag)
-        );
+        const currentWar = await this.clanWarService.getCurrentByTag(clanTag);
         await this.bandService.refreshPost(currentWar);
     };
 
     checkStatus = async (clanTag: ClanTag) => {
-        const currentWar = new CurrentWar(
-            await this.cocApi.getClanWarByTag(clanTag)
-        );
+        const currentWar = await this.clanWarService.getCurrentByTag(clanTag);
         await new LineNotify().sendMessage(currentWar.state);
     };
 }
